ensure ChangeProfilePicture not call DeleteFile if file does not exists and SaveUserPicture throws

+++ src/domain/usecases/change-profile-picture.ts
@@ -24,7 +24,11 @@ export const setupChangeProfilePicture: Setup = (fileStorage, idGenerator, userP
   try {
     await userProfileRepo.savePicture(userProfile)
   } catch {
-    await fileStorage.delete({ key: uniqueId })
+    if (file !== undefined) {
+      await fileStorage.delete({ key: uniqueId })
+    }
+
+    throw new Error()
   }
 
   return {
